File Structure Update
