+++ src/api/ding/ding.ts
@@ -1,0 +1,6 @@
+import Elysia from "elysia";
+import dingGET from "./get";
+
+const ding = new Elysia({prefix: "/ding"}).use(dingGET)
+
+export default ding
