+++ src/api/ding/get.tsx
@@ -1,0 +1,7 @@
+import Elysia from "elysia";
+
+const dingGET = new Elysia().get("/", async () => {
+  return <p>Ding! Ringed Server!</p>;
+});
+
+export default dingGET
