+++ src/api/todos/toggle/toggle.ts
@@ -1,0 +1,6 @@
+import Elysia from "elysia";
+import todoTogglePOST from "./post";
+
+const todosToggle = new Elysia({ prefix: "/todos/toggle" }).use(todoTogglePOST);
+
+export default todosToggle;
